backend/requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
Pillow>=9.0.0
python-dotenv>=1.0.0
requests>=2.28.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
bcrypt>=4.0.0
